requirements.txt
feedparser
beautifulsoup4
requests
anthropic
pyyaml
schedule
